+++ server/src/fixes.d.ts
@@ -1,0 +1,4 @@
+declare module "engine.io-parser";
+
+declare type RawData = {};
+declare type Packet = {};